[python-frontend] hoist shared reducer helpers, fix axis=None handling
Code review (Copilot) on the axis-aware reducer commit found that
numpy_reducer_has_unsupported_keywords() rejects axis=None: once the
axis-aware fast path declines an axis (returning nullopt specifically
for None, meaning "flatten"), this later, unconditional "any keyword
at all is unsupported" check still sees the axis= keyword sitting in
the call's own AST and rejects it. Fixed by making it tolerate a lone
axis keyword whose value is the JSON None literal specifically (any
other value cannot reach this point unrejected).

Tracing this further found the true root cause sits one level up:
extract_literal_axis_keyword() -- used by every one of these reducers'
own axis parsing -- checked axis_kw->is_null() to detect an absent/
None axis, but axis=None is a Constant node whose *value* is JSON
null, not a bare JSON null at that position. That check never matched,
so it fell into the literal-axis extraction and threw "requires a
literal axis" before ever reaching the (now-fixed) keyword check
above -- meaning the first fix was unreachable until this one landed
too. Both now share the same is_json_none_literal() detection.

Lucas Cordeiro's review on the same commit noted numpy_call_expr.cpp
and function_call/expr.cpp each carry a byte-for-byte duplicate of
build_1d_numpy_array_value()/normalize_reducer_axis(), and
function_call/expr.cpp hand-rolls its own narrower axis-literal parser
instead of reusing try_extract_numeric_constant(). Pulled numeric_value,
try_extract_numeric_constant(), is_json_none_literal(),
build_1d_numpy_array_value() and normalize_reducer_axis() into a new
numpy_reducer_shared.h so both translation units share one copy.
